FILE: apc-aa/admin/index.php3 Changed by: Setu Change:
New options in Query Strings to hide few items from the page.
By default, it shows everything as original version.

navbar=0 - to hide the navigation bar at the top of page.
leftbar=0 - to hide the navigation bar at the left of page.
footer=0 - to hide the footer message
sort_filter=0 - to hide box for sort / filter.
action_selected=0 - to hide box for the actions for selected (marked) item

Big switch to hide many things, to show only the title and item list.
bodyonly=1 - this single switch will hides all of the items above.